--- lua/plugins/completion/init.lua
@@ -1,3 +1,0 @@
-return {
-	require("plugins.completion.cmp"),
-}
