--- lua/plugins/editing/init.lua
@@ -1,9 +1,0 @@
-return {
-	require("plugins.editing.textobjects"),
-	require("plugins.editing.pairs"),
-	require("plugins.editing.comments"),
-	require("plugins.editing.zenmode"),
-	require("plugins.editing.whichkey"),
-	require("plugins.editing.cheatsheet"),
-	-- { "mbbill/undotree" },
-}
