--- lua/plugins/editing/textobjects.lua
@@ -1,18 +1,0 @@
-return {
-	{
-		"chrisgrieser/nvim-various-textobjs",
-		opts = { useDefaultKeymaps = true },
-	},
-	{
-		"windwp/nvim-ts-autotag", -- For automatically closing HTML tags
-		dependencies = { "nvim-treesitter/nvim-treesitter" },
-		opts = {
-			autotag = {
-				enable = true,
-				enable_rename = true,
-				enable_close = true,
-				enable_close_on_slash = false,
-			},
-		},
-	},
-}
