+++ lua/plugins/harpoon.lua
@@ -1,0 +1,7 @@
+return {
+	{
+		"ThePrimeagen/harpoon",
+		branch = "harpoon2",
+		dependencies = { "nvim-lua/plenary.nvim" },
+	},
+}
