+++ lua/plugins/icons.lua
