+++ lua/plugins/mini-icons.lua
@@ -1,0 +1,12 @@
+return {
+	{
+		"echasnovski/mini.icons",
+		dependencies = {
+			"nvim-tree/nvim-web-devicons",
+		},
+		config = function()
+			require("mini.icons").setup()
+			-- MiniIcons.mock_nvim_web_devicons()
+		end,
+	},
+}
