+++ lua/plugins/numb.lua
@@ -1,0 +1,12 @@
+return {
+	{
+		"nacro90/numb.nvim",
+		opts = {
+			show_numbers = true, -- Enable 'number' for the window while peeking
+			show_cursorline = true, -- Enable 'cursorline' for the window while peeking
+			hide_relativenumbers = true, -- Enable turning off 'relativenumber' for the window while peeking
+			number_only = true, -- Peek only when the command is only a number instead of when it starts with a number
+			centered_peeking = true, -- Peeked line will be centered relative to window
+		},
+	},
+}
