--- lua/plugins/search/init.lua
@@ -1,4 +1,0 @@
-return {
-	require("plugins.search.telescope"),
-	require("plugins.search.spectre"),
-}
