+++ lua/plugins/spectre.lua
