--- lua/plugins/terminal/init.lua
@@ -1,3 +1,0 @@
-return {
-	require("plugins.terminal.toggleterm"),
-}
