+++ lua/plugins/textobjects.lua
@@ -1,0 +1,3 @@
+return {
+	{ "chrisgrieser/nvim-various-textobjs", opts = { useDefaultKeymaps = true } },
+}
