+++ lua/plugins/trouble.lua
@@ -7,9 +7,9 @@ return {
 			follow = false,
 			-- auto_preview = false, -- automatically open preview when on an item
 			auto_refresh = false, -- auto refresh when open
-            win = {
-                -- position = "right"
-            }
+			win = {
+				-- position = "right"
+			},
 		},
 	},
 }
