--- lua/plugins/ui/init.lua
@@ -1,9 +1,0 @@
-return {
-	-- require("plugins.ui.breadcrumb"),
-	require("plugins.ui.indent"),
-	require("plugins.ui.startup"),
-	require("plugins.ui.statusline").plugin,
-	require("plugins.ui.tabs"),
-	{ "RRethy/vim-illuminate" },
-	require("plugins.ui.icons"),
-}
